+++ src/backends/sentry_backend_none.c
@@ -1,5 +1,10 @@
 #include "sentry_backend.h"
 
+void
+sentry__backend_preload(void)
+{
+}
+
 sentry_backend_t *
 sentry__backend_new(void)
 {